NewCharacterPriorityTable: Changed the default settings while testing the new skills tab.
SwipeFragmentHolder: Moved the setContentView to the bottom of OnCreate so the ShadowrunCharacter gets a chance to generate.

Skill: Added Specialization array needed for dropdown menu.

AttributeFragement: Set min width display on the attribute textview, Changed karmaCounter to karmaUnsued so that it made more sense.

QualitiesFragment: Removed an unnecessary  comment.

Skill/MainContainer: Added the use of skill counters upon character creation, and karma use.

ChummerConstants: Changed the Tag name so that it is more unique